+++ nomad/datasource_nomad_job_test.go
@@ -2,49 +2,66 @@ package nomad
 
 import (
 	"fmt"
-	"regexp"
 	"testing"
+
+	"github.com/hashicorp/nomad/api"
 
 	"github.com/hashicorp/terraform/helper/resource"
 	"github.com/hashicorp/terraform/terraform"
 )
 
 func TestAccDataSourceNomadJob_Basic(t *testing.T) {
+	job := "testjobds"
 	resource.Test(t, resource.TestCase{
 		PreCheck:     func() { testAccPreCheck(t) },
 		Providers:    testProviders,
-		CheckDestroy: testResourceJob_forceDestroyWithPurge("foo"),
+		CheckDestroy: testResourceJob_forceDestroyWithPurge(job, "default"),
 		Steps: []resource.TestStep{
 			{
-				Config: testAccCheckDataSourceNomadJobConfig,
-			},
-			{
-				Config: testAccCheckDataSourceNomadJobConfigWithJob,
-				Check:  testAccCheckDataSourceNomadJobExists("data.nomad_job.foobaz"),
-			},
-			{
-				Config: testAccCheckDataSourceNomadJobConfigWithJob,
+				Config: testAccJobDataSourceConfig(job),
 				Check: resource.ComposeTestCheckFunc(
+					testAccDataSourceNomadJobExists("data.nomad_job.test-job", "default"),
 					resource.TestCheckResourceAttr(
-						"data.nomad_job.foobaz", "name", "foo"),
+						"data.nomad_job.test-job", "name", job),
 					resource.TestCheckResourceAttr(
-						"data.nomad_job.foobaz", "type", "service"),
+						"data.nomad_job.test-job", "type", "batch"),
 					resource.TestCheckResourceAttr(
-						"data.nomad_job.foobaz", "status", "running"),
+						"data.nomad_job.test-job", "priority", "50"),
 					resource.TestCheckResourceAttr(
-						"data.nomad_job.foobaz", "priority", "50"),
+						"data.nomad_job.test-job", "namespace", "default"),
 				),
-			},
-			{
-				Config:      testAccCheckDataSourceNomadJobConfigErr,
-				Destroy:     false,
-				ExpectError: regexp.MustCompile(`.*job not found`),
 			},
 		},
 	})
 }
 
-func testAccCheckDataSourceNomadJobExists(n string) resource.TestCheckFunc {
+func TestAccDataSourceNomadJob_Namespaced(t *testing.T) {
+	ns := "jobds-test-namespace"
+	job := "testjobds-namespace"
+	resource.Test(t, resource.TestCase{
+		PreCheck:     func() { testAccPreCheck(t); testCheckPro(t) },
+		Providers:    testProviders,
+		CheckDestroy: testResourceJob_forceDestroyWithPurge(job, ns),
+		Steps: []resource.TestStep{
+			{
+				Config: testAccNSJobDataSourceConfig(job, ns),
+				Check: resource.ComposeTestCheckFunc(
+					testAccDataSourceNomadJobExists("data.nomad_job.test-job", ns),
+					resource.TestCheckResourceAttr(
+						"data.nomad_job.test-job", "name", job),
+					resource.TestCheckResourceAttr(
+						"data.nomad_job.test-job", "type", "batch"),
+					resource.TestCheckResourceAttr(
+						"data.nomad_job.test-job", "priority", "50"),
+					resource.TestCheckResourceAttr(
+						"data.nomad_job.test-job", "namespace", ns),
+				),
+			},
+		},
+	})
+}
+
+func testAccDataSourceNomadJobExists(n, namespace string) resource.TestCheckFunc {
 	return func(s *terraform.State) error {
 		rs, ok := s.RootModule().Resources[n]
 		if !ok {
@@ -61,7 +78,9 @@ func testAccCheckDataSourceNomadJobExists(n string) resource.TestCheckFunc {
 		id := rs.Primary.ID
 
 		// Try to find the job
-		test_job, _, err := client.Jobs().Info(id, nil)
+		test_job, _, err := client.Jobs().Info(id, &api.QueryOptions{
+			Namespace: namespace,
+		})
 
 		if err != nil {
 			return err
@@ -75,40 +94,24 @@ func testAccCheckDataSourceNomadJobExists(n string) resource.TestCheckFunc {
 	}
 }
 
-func testAccCheckNomadJobDestroy(s *terraform.State) error {
-	providerConfig := testProvider.Meta().(ProviderConfig)
-	client := providerConfig.client
-
-	for _, rs := range s.RootModule().Resources {
-		if rs.Type != "nomad_job" {
-			continue
-		}
-
-		id := rs.Primary.ID
-
-		_, _, err := client.Jobs().Deregister(id, false, nil)
-		if err != nil {
-			return fmt.Errorf("error deregistering job: %s", err)
-		}
-	}
-
-	return nil
+func testAccNSJobDataSourceConfig(job, ns string) string {
+	return `
+resource "nomad_namespace" "ns-instance" {
+  name = "` + ns + `" 
 }
 
-var testAccCheckDataSourceNomadJobConfig = `
-resource "nomad_job" "foobar" {
+resource "nomad_job" "job-instance" {
 	jobspec = <<EOT
-		job "foo" {
+		job "` + job + `" {
 			datacenters = ["dc1"]
-			type = "service"
+			type = "batch"
+			namespace = "${nomad_namespace.ns-instance.name}"
 			group "foo" {
 				task "foo" {
-					leader = true ## new in Nomad 0.5.6
-
 					driver = "raw_exec"
 					config {
-						command = "/bin/sleep"
-						args = ["1"]
+						command = "/bin/echo"
+						args = ["test"]
 					}
 
 					resources {
@@ -125,16 +128,46 @@ resource "nomad_job" "foobar" {
 		}
 	EOT
 }
-`
 
-var testAccCheckDataSourceNomadJobConfigWithJob = testAccCheckDataSourceNomadJobConfig + `
-data "nomad_job" "foobaz" {
-  job_id               = "foo"
+data "nomad_job" "test-job" {
+  job_id    = "${nomad_job.job-instance.id}"
+  namespace = "${nomad_job.job-instance.namespace}"
 }
 `
+}
 
-var testAccCheckDataSourceNomadJobConfigErr = testAccCheckDataSourceNomadJobConfig + `
-data "nomad_job" "foobar" {
-  job_id               = "foo-missing"
+func testAccJobDataSourceConfig(job string) string {
+	return `
+resource "nomad_job" "job-instance" {
+	jobspec = <<EOT
+		job "` + job + `" {
+			datacenters = ["dc1"]
+			type = "batch"
+			group "foo" {
+				task "foo" {
+					driver = "raw_exec"
+					config {
+						command = "/bin/echo"
+						args = ["test"]
+					}
+
+					resources {
+						cpu = 100
+						memory = 10
+					}
+
+					logs {
+						max_files = 3
+						max_file_size = 10
+					}
+				}
+			}
+		}
+	EOT
+}
+
+data "nomad_job" "test-job" {
+  job_id    = "${nomad_job.job-instance.id}"
 }
 `
+}