+++ nomad/resource_namespace.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"log"
 	"strings"
+	"time"
 
 	"github.com/hashicorp/nomad/api"
 	"github.com/hashicorp/terraform/helper/schema"
@@ -69,9 +70,22 @@ func resourceNamespaceDelete(d *schema.ResourceData, meta interface{}) error {
 	name := d.Id()
 
 	log.Printf("[DEBUG] Deleting namespace %q", name)
-	_, err := client.Namespaces().Delete(name, nil)
-	if err != nil {
-		return fmt.Errorf("error deleting namespace %q: %s", name, err.Error())
+	retries := 0
+	for {
+		_, err := client.Namespaces().Delete(name, nil)
+		if err == nil {
+			break
+		} else if retries < 10 {
+			if strings.Contains(err.Error(), "has non-terminal jobs") {
+				log.Printf("[WARN] could not delete namespace %q because of non-terminal jobs, will pause and retry", name)
+				time.Sleep(5 * time.Second)
+				retries++
+				continue
+			}
+			return fmt.Errorf("error deleting namespace %q: %s", name, err.Error())
+		} else {
+			return fmt.Errorf("too many failures attempting to delete namespace %q: %s", name, err.Error())
+		}
 	}
 	log.Printf("[DEBUG] Deleted namespace %q", name)